playground-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
transformers==4.36.2
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
librosa==0.10.1
numpy==1.24.4
python-multipart==0.0.6